Wombling algorithm
Also includes additional code to view the densities, trace, posterior mean for W, acceptance rates for W and initial W.
Additional code at the bottom also includes option to process chains, and apply a burn-in and thinning of the MCMC chains, and plot diagnostics again. And finally save processed output to pdf.